app/components are covered.
