+++ .jest.js
@@ -1,5 +1,10 @@
 module.exports = {
-	verbose: true,
+	verbose: false,
+	silent: true,
+	setupFiles: ['./testSetup.js'],
+	moduleNameMapper: {
+		"\\.(css|svg)": "identity-obj-proxy"
+	}
 
 	// collectCoverage: true,
 	// coverageDirectory: '.coverage',
