+++ package.json
@@ -31,11 +31,15 @@
     "babel-preset-stage-0": "^6.24.1",
     "cross-env": "^5.1.1",
     "css-loader": "^0.28.7",
+    "enzyme": "^3.2.0",
+    "enzyme-adapter-react-16": "^1.1.0",
     "eslint": "^4.11.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "flow-bin": "^0.59.0",
     "flow-runtime": "^0.16.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^21.3.0-beta.8",
+    "jest-css-modules": "^1.1.0",
     "json-loader": "^0.5.7",
     "postcss": "^6.0.14",
     "postcss-loader": "^2.0.8",
