+++ source/app/App.js
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import Topbar from '@app/components/Topbar';
+import { Topbar } from '@app/components/Topbar';
 import { ActionBar } from '@app/components/ActionBar';
 import { TestsList } from '@app/scenes/TestsList';
 import "./styles/App.css";
