+++ source/app/components/ActionBar/_tests/ActionBar.spec.js
@@ -1,0 +1,17 @@
+import * as React from 'react';
+import { shallow } from 'enzyme';
+
+import ActionBar from '../ActionBar';
+
+const mockStore = {
+	recording: false,
+	actions: {
+		toggleRecording() { }
+	}
+}
+
+test('renders correctly', () => {
+	const component = shallow(<ActionBar $tests={mockStore} />);
+
+	expect(component).toBeTruthy();
+})
