+++ source/app/components/ErrorBoundary/_tests/ErrorBoundary.spec.js
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { shallow } from 'enzyme';
+
+import ErrorBoundary from '../ErrorBoundary';
+
+test('renders correctly', () => {
+	const component = shallow(<ErrorBoundary />);
+
+	expect(component).toBeTruthy();
+})
