+++ source/app/components/EventTypeBox/EventTypeBox.js
@@ -1,34 +1,24 @@
 import * as React from "react";
 
-import ErrorBoundary from '@app/components/ErrorBoundary';
-import Icon from '@app/components/Icon';
+import { ErrorBoundary } from '@app/components/ErrorBoundary';
 import KeyboardIcon from '@app/assets/svgs/keyboard.svg'
 import PointerIcon from '@app/assets/svgs/mouse-arrow.svg'
 import YieldIcon from '@app/assets/svgs/yield.svg'
 
 import "./styles/EventTypeBox.css";
 
-const colors = {
-	mouse: "red",
-	keyboard: "green",
-	assertion: "blue"
-};
-
 interface PropTypes {
 	type: string
 }
 
 const EventTypeBox = (props: PropTypes) => {
-	const color = colors[props.type];
-
-
-	const { Icon, size } = do {
+	const { Icon, size, color } = do {
 		if (props.type === "mouse") {
-			({ Icon: PointerIcon, size: 26 })
+			({ Icon: PointerIcon, size: 26, color: 'red' })
 		} else if (props.type === "keyboard") {
-			({ Icon: KeyboardIcon, size: 32 })
+			({ Icon: KeyboardIcon, size: 32, color: 'blue' })
 		} else if (props.type === "assertion") {
-			({ Icon: YieldIcon, size: 28 })
+			({ Icon: YieldIcon, size: 28, color: 'green' })
 		}
 	};
